style : 디테일페이지 스타일링
