+++ db.json
@@ -19,8 +19,8 @@
     {
       "id": "3",
       "userId": "heerok",
-      "title": "윤시커피로스터스",
-      "postContent": "022 IMAB 에스프레소 콩쿠르 우승 2022 IMAB 에스프레소 콩쿠르 우승 2022 IMAR 로스팅 콩쿠르 5위한 커피 집",
+      "title": "커피맛집 찾았어요!",
+      "postContent": "에스프레소 콩쿠르 우승한 집이래요! 어쩐지 맛있더라 다들 한번 맛보러 가세요!",
       "foodType": "카페",
       "address": "경기 고양시 일산동구 일산로316번길 25-9 1층 101호"
     }
